Remove old comments and add Javadoc comments.
