+++ backend/src/api/countries.js
@@ -2,7 +2,9 @@ const express = require('express');
 const router = express.Router();
 const pool = require('../db');
 
-// Ensure overrides table exists (idempotent)
+/**
+ * Ensures the overrides table exists; safe to call repeatedly.
+ */
 async function ensureOverridesTable() {
   await pool.query(`
     CREATE SCHEMA IF NOT EXISTS gppd;
@@ -19,10 +21,8 @@ router.get('/stats/top', async (req, res) => {
   try {
     const { limit = 10 } = req.query;
 
-    // Ensure overrides table exists so LEFT JOIN works even if empty
     await ensureOverridesTable();
 
-    // Compute base aggregates per country from plants, then merge capacity overrides
     const result = await pool.query(
       `
       WITH base AS (
