+++ backend/tests/helpers/setupTestEnv.js
@@ -1,7 +1,6 @@
 const path = require('path');
 require('dotenv').config({ path: path.resolve(__dirname, '../../../.env.test') });
 
-// Default sensible test env fallbacks if .env.test isn't provided
 process.env.API_KEY = process.env.API_KEY || 'test-api-key';
 process.env.DB_HOST = process.env.DB_HOST || 'localhost';
 process.env.DB_PORT = process.env.DB_PORT || '5432';
