+++ frontend/components/KpiStats.jsx
@@ -4,6 +4,9 @@ import { useEffect, useMemo, useState } from 'react'
 import { Card, CardContent } from '@/components/ui/card'
 import { Zap, BarChart3, Factory } from 'lucide-react'
 
+/**
+ * Displays high-level KPIs for either global or single-country mode.
+ */
 export default function KpiStats({ selectedCountries = [] }) {
   const [top, setTop] = useState(null)
   const [fuels, setFuels] = useState(null)
